(renamed) rate count 변경 / URL 변경

+++ src/main/java/today/creame/web/influence/application/model/InfluenceResult.java
@@ -19,7 +19,7 @@ public class InfluenceResult {
     private String extensionNumber;
     private String nickname;
     private float rate;
-    private float rateCount;
+    private int rateCount;
     private int reviewCount;
     private int qnaCount;
